requirements.txt
mock==4.0.2
matplotlib==3.3.2
lime==0.2.0.1
numpy==1.19.2
plac==1.2.0
scipy==1.5.2
pandas==1.1.2
joblib==0.17.0
PyYAML==5.3.1
scikit_learn==0.23.2
dvc<2
black==20.8b1